Separate tests from  implementation.
