+++ src/main.rs
@@ -6,6 +6,9 @@ use std::path::Path;
 
 use clap::{App, Arg};
 
+#[cfg(test)]
+mod tests;
+
 // TODO{code}: improve program flow; especially regarding errors/panics
 // TODO{test}: improve testing
 // TODO{code,test}: test on constraints of ICO format (256x256) before trying to convert
@@ -92,16 +95,14 @@ fn main() {
     // TODO{docs}: supported input formats of the image crate.
 
     let matches = App::new("Simple Image Converter")
-        .version("0.2.0")
+        .version("0.3.0")
         .author("foresterre <garm@ilumeo.com>")
         .about("Converts an image from one format to another.\n\n\
-                Supported input formats are PNG, JPEG (baseline, progressive), GIF, BMP\
-                ICO, TIFF (Baseline(no fax support) + LZW + PackBits), Webp (Lossy(Luma channel only))\
-                PNM (PBM, PGM, PPM, standard PAM) [1] \n\n\
-                The image conversion is actually done by the awesome 'image' crate. \n\
-                Sic is only a command line frontend which supports a part of the \
+                Supported input formats are described BMP, GIF, ICO, JPEG, PNG, PPM (limitations may apply). \n\n\
+                The image conversion is actually done by the awesome 'image' crate [1]. \n\
+                Sic itself is a small command line frontend which supports a small part of the \
                 conversion operations supported by the 'image' library. \n\n\
-                [1] source: https://github.com/PistonDevelopers/image#2-supported-image-formats \n\n\
+                [1] image crate by PistonDevelopers: https://github.com/PistonDevelopers/image \n\n\
                 ")
         .arg(Arg::with_name("FORCED_OUTPUT_FORMAT")
             .short("f")
@@ -138,159 +139,4 @@ fn main() {
     if let Some(format) = final_format {
         convert_image(in_file, out_file, format)
     }
-
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // which_image_format
-
-    // GIF
-
-    #[test]
-    fn image_format_from_str_gif_uc() {
-        assert_eq!(Some(image::GIF), image_format_from_str("GIF"));
-    }
-
-    #[test]
-    fn image_format_from_str_gif_lc() {
-        assert_eq!(Some(image::GIF), image_format_from_str("gif"));
-    }
-
-    #[test]
-    fn image_format_from_str_gif_mc() {
-        assert_eq!(Some(image::GIF), image_format_from_str("gIF"));
-    }
-
-    // ICO
-
-    #[test]
-    fn image_format_from_str_ico_uc() {
-        assert_eq!(Some(image::ICO), image_format_from_str("ICO"));
-    }
-
-    #[test]
-    fn image_format_from_str_ico_lc() {
-        assert_eq!(Some(image::ICO), image_format_from_str("ico"));
-    }
-
-    #[test]
-    fn image_format_from_str_ico_mc() {
-        assert_eq!(Some(image::ICO), image_format_from_str("icO"));
-    }
-
-    // JPG/JPEG
-
-    #[test]
-    fn image_format_from_str_jpeg_uc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("JPEG"));
-    }
-
-    #[test]
-    fn image_format_from_str_jpeg_lc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("jpeg"));
-    }
-
-    #[test]
-    fn image_format_from_str_jpeg_mc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("jPeG"));
-    }
-
-    #[test]
-    fn image_format_from_str_jpg_uc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("JPG"));
-    }
-
-    #[test]
-    fn image_format_from_str_jpg_lc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("jpg"));
-    }
-
-    #[test]
-    fn image_format_from_str_jpg_mc() {
-        assert_eq!(Some(image::JPEG), image_format_from_str("jPG"));
-    }
-
-
-    // PNG
-
-    #[test]
-    fn image_format_from_str_png_uc() {
-        assert_eq!(Some(image::PNG), image_format_from_str("PNG"));
-    }
-
-    #[test]
-    fn image_format_from_str_png_lc() {
-        assert_eq!(Some(image::PNG), image_format_from_str("png"));
-    }
-
-    #[test]
-    fn image_format_from_str_png_mc() {
-        assert_eq!(Some(image::PNG), image_format_from_str("pNg"));
-    }
-
-    // PPM
-
-    #[test]
-    fn image_format_from_str_ppm_uc() {
-        assert_eq!(Some(image::PPM), image_format_from_str("PPM"));
-    }
-
-    #[test]
-    fn image_format_from_str_ppm_lc() {
-        assert_eq!(Some(image::PPM), image_format_from_str("ppm"));
-    }
-
-    #[test]
-    fn image_format_from_str_ppm_mc() {
-        assert_eq!(Some(image::PPM), image_format_from_str("pPm"));
-    }
-
-    // determine_format_by_extension
-
-    #[test]
-    fn determine_format_by_extension_ok_path() {
-        assert_eq!(Some(image::PNG), determine_format_by_extension("C:/users/some/path.png"));
-    }
-
-    #[test]
-    fn determine_format_by_extension_test_ok_file() {
-        assert_eq!(Some(image::PNG), determine_format_by_extension("path.png"));
-    }
-
-    #[test]
-    fn determine_format_by_extension_test_no_ext_path() {
-        assert_eq!(None, determine_format_by_extension("C:/users/some/png"));
-    }
-
-    #[test]
-    fn determine_format_by_extension_test_no_ext_file() {
-        assert_eq!(None, determine_format_by_extension("png"));
-    }
-
-    // get_extension
-
-    #[test]
-    fn get_extension_ok_path() {
-        assert_eq!(Some("png"), get_extension("C:/users/some/path.png"));
-    }
-
-    #[test]
-    fn get_extension_test_ok_file() {
-        assert_eq!(Some("png"), get_extension("path.png"));
-    }
-
-    #[test]
-    fn get_extension_test_no_ext_path() {
-        assert_eq!(None, get_extension("C:/users/some/png"));
-    }
-
-    #[test]
-    fn get_extension_test_no_ext_file() {
-        assert_eq!(None, get_extension("png"));
-    }
-
-}
-
